requirements.txt
Flask
Werkzeug
requests
ffmpeg-python
openai-whisper
gunicorn
APScheduler
srt
numpy
torch
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

